requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-dotenv==1.0.0
openai==0.27.6
pydantic==1.10.0
requests==2.28.2
beautifulsoup4==4.11.2
pdfkit
jinja2
wkhtmltopdf
certifi